# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API documentation for the Library application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "List all loans",
                "description": "Returns a list of loans",
                "operationId": "7a6d018dd0b622136235484db1036de8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Loan"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Create a new loan",
                "description": "Creates a new loan",
                "operationId": "1f6c6e0db1616c54e249195c256856c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Loan"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/loans/{id}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Get a loan by ID",
                "description": "Returns a single loan",
                "operationId": "a4a657443e9563396f74613aa1059203",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the loan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Loans"
                ],
                "summary": "Update an existing loan",
                "description": "Updates an existing loan",
                "operationId": "9851fce80ce3bb9682a7ae6d34fda306",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the loan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Loan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loan updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Loans"
                ],
                "summary": "Delete a loan",
                "description": "Deletes a loan by ID",
                "operationId": "3aaa9683b74a4086d39b53b57c0598d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the loan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Loan deleted"
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "properties": {
                    "id": {
                        "description": "ID of the author",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the author",
                        "type": "string"
                    },
                    "birth_date": {
                        "description": "Birth date of the author",
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "description": "Timestamp when the author was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the author was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Book": {
                "properties": {
                    "id": {
                        "description": "ID of the book",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string"
                    },
                    "publication_year": {
                        "description": "Publication year of the book",
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "description": "Timestamp when the book was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the book was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "authors": {
                        "description": "List of authors associated with the book",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    }
                },
                "type": "object"
            },
            "Loan": {
                "properties": {
                    "id": {
                        "description": "ID of the loan",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID of the user who borrowed the book",
                        "type": "integer",
                        "format": "int64"
                    },
                    "book_id": {
                        "description": "ID of the borrowed book",
                        "type": "integer",
                        "format": "int64"
                    },
                    "borrowed_at": {
                        "description": "Date when the book was borrowed",
                        "type": "string",
                        "format": "date-time"
                    },
                    "returned_at": {
                        "description": "Date when the book was returned",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Timestamp when the loan record was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the loan record was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "book": {
                        "$ref": "#/components/schemas/Book"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the user was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the user was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}